Restructured frontend app and made login look similar to app layout
