+++ frontend/src/App.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { BrowserRouter, Routes, Route } from "react-router-dom";
-import Login from './login';
 import Layout from './components/Layout';
+import Login from './pages/Login';
 import Home from './pages/Home';
 
 import "./styles/tailwind.css";
@@ -10,9 +10,9 @@ function App() {
   return (
     <BrowserRouter>
       <Routes>
+        <Route path="/login" element={<Login />}/>
         <Route path="/" element={<Layout />} >
           <Route index element={<Home />} />
-          <Route path="login" element={<Login />} />
         </Route>
       </Routes>
     </BrowserRouter>
