--- frontend/src/login.css
@@ -1,81 +1,0 @@
-.wrapper {
-	zoom: 1;
-	overflow: hidden;
-	font-size : 17px;
-	min-width : 360px;
-	margin : 0 auto;
-	background-color: #F5F5F5;
-}
-.logo {
-	vertical-align : middle;
-	width : 32px;
-	height : 32px;
-}
-.Header {
-	width : 300px;
-	text-align : center;
-	margin : 0 auto;
-	margin-top : 30px;
-	margin-bottom : 25px;
-	background-color : #F5F5F5;
-}
-.Header h1 {
-	display : inline-block;
-}
-.Header h1 span {
-	vertical-align: middle;
-}
-
-.Login {
-	width : 350px;
-	margin : 0 auto;
-	border-radius : 10px;
-	background-color : #FFF;
-	padding : 20px 0;
-	margin-bottom : 10px;
-}
-.form {
-	text-align : center;
-}
-
-.form input {
-	padding : 0 20px;
-	background: rgba(79,79,79,0.21);
-    border: rgba(79,79,79,0.21);
-    border-radius: 5px;
-	width : 90%;
-	height : 40px;
-	margin-bottom : 10px;
-}
-.form submit {
-	display : inline-block;
-	background-color : #2562FF;
-	color : white;
-	border-radius: 5px;
-	width : 90%;
-	line-height : 40px;
-}
-
-.signup {
-	margin-top : 10px;
-	display : inline-block;
-	background-color : #FF8E25;
-	color : white;
-	border-radius: 5px;
-	width : 90%;
-	line-height : 40px;
-}
-
-
-
-
-
-.Introduce {
-	width : 350px;
-	height : 300px;
-	margin : 0 auto;
-	border-radius : 10px;
-	background-color : #FFF;
-	padding : 20px 0;
-	margin-bottom : 10px;
-}
